Moved dashboard to pages
