+++ apps/web/src/main.tsx
@@ -13,10 +13,13 @@ import Task from "./pages/Task/Task.tsx";
 import AuthGuard from "./auth/AuthGuard.tsx";
 import Pricing from "./pages/Pricing/Pricing.tsx";
 import store from "./store/store";
-import Dashboard from "./components/Dashboard/Dashboard.tsx";
+import Dashboard from "./pages/Dashboard/Dashboard.tsx";
 
 const oidcConfig: AuthProviderProps = {
-  authority: import.meta.env.VITE_KEYCLOAK_URL + "/realms/" + import.meta.env.VITE_KEYCLOAK_REALM,
+  authority:
+    import.meta.env.VITE_KEYCLOAK_URL +
+    "/realms/" +
+    import.meta.env.VITE_KEYCLOAK_REALM,
   client_id: import.meta.env.VITE_KEYCLOAK_CLIENT_ID,
   redirect_uri: import.meta.env.VITE_KEYCLOAK_REDIRECT_URI,
   post_logout_redirect_uri: import.meta.env
